refactor helpers database
