+++ client/src/androidUnitTest/kotlin/org/apptank/horus/client/sync/manager/SynchronizatorManagerTest.kt
@@ -4,11 +4,11 @@ import org.apptank.horus.client.TestCase
 import org.apptank.horus.client.auth.HorusAuthentication
 import org.apptank.horus.client.base.DataMap
 import org.apptank.horus.client.base.DataResult
-import org.apptank.horus.client.control.ISyncControlDatabaseHelper
+import org.apptank.horus.client.control.helper.ISyncControlDatabaseHelper
 import org.apptank.horus.client.control.SyncControl
 import org.apptank.horus.client.data.Horus
-import org.apptank.horus.client.database.DatabaseOperation
-import org.apptank.horus.client.database.IOperationDatabaseHelper
+import org.apptank.horus.client.database.struct.DatabaseOperation
+import org.apptank.horus.client.control.helper.IOperationDatabaseHelper
 import org.apptank.horus.client.di.INetworkValidator
 import org.apptank.horus.client.sync.network.dto.SyncDTO
 import org.apptank.horus.client.sync.network.service.ISynchronizationService
