+++ client/src/commonMain/kotlin/org/apptank/horus/client/HorusDataFacade.kt
@@ -6,14 +6,14 @@ import org.apptank.horus.client.base.CallbackEvent
 import org.apptank.horus.client.base.CallbackNullable
 import org.apptank.horus.client.base.DataMap
 import org.apptank.horus.client.base.DataResult
-import org.apptank.horus.client.control.ISyncControlDatabaseHelper
+import org.apptank.horus.client.control.helper.ISyncControlDatabaseHelper
 import org.apptank.horus.client.data.DataChangeListener
 import org.apptank.horus.client.data.Horus
-import org.apptank.horus.client.database.DatabaseOperation
-import org.apptank.horus.client.database.IOperationDatabaseHelper
-import org.apptank.horus.client.database.SQL
+import org.apptank.horus.client.database.struct.DatabaseOperation
+import org.apptank.horus.client.control.helper.IOperationDatabaseHelper
+import org.apptank.horus.client.database.struct.SQL
 import org.apptank.horus.client.database.builder.SimpleQueryBuilder
-import org.apptank.horus.client.database.mapToDBColumValue
+import org.apptank.horus.client.database.struct.mapToDBColumValue
 import org.apptank.horus.client.di.HorusContainer
 import org.apptank.horus.client.di.INetworkValidator
 import org.apptank.horus.client.eventbus.EventBus
