+++ client/src/commonMain/kotlin/org/apptank/horus/client/control/helper/ISyncFileDatabaseHelper.kt
@@ -1,0 +1,5 @@
+package org.apptank.horus.client.control.helper
+
+interface ISyncFileDatabaseHelper {
+
+}
