+++ client/src/commonMain/kotlin/org/apptank/horus/client/database/SyncFileDatabaseHelper.kt
@@ -1,0 +1,12 @@
+package org.apptank.horus.client.database
+
+import app.cash.sqldelight.db.SqlDriver
+import org.apptank.horus.client.control.helper.ISyncFileDatabaseHelper
+
+class SyncFileDatabaseHelper(
+    databaseName: String,
+    driver: SqlDriver,
+) : SQLiteHelper(driver, databaseName), ISyncFileDatabaseHelper {
+
+
+}
