+++ client/src/commonMain/kotlin/org/apptank/horus/client/database/struct/DatabaseOperation.kt
@@ -1,4 +1,4 @@
-package org.apptank.horus.client.database
+package org.apptank.horus.client.database.struct
 
 import org.apptank.horus.client.control.SyncControl
 import org.apptank.horus.client.data.Horus
