+++ client/src/commonMain/kotlin/org/apptank/horus/client/sync/upload/UploadFileRepository.kt
@@ -1,0 +1,10 @@
+package org.apptank.horus.client.sync.upload
+
+import org.apptank.horus.client.sync.network.service.IFileSynchronizationService
+
+class UploadFileRepository(
+    private val service: IFileSynchronizationService
+) {
+
+
+}
