+++ client/src/commonMain/kotlin/org/apptank/horus/client/tasks/SynchronizeDataTask.kt
@@ -1,7 +1,7 @@
 package org.apptank.horus.client.tasks
 
-import org.apptank.horus.client.control.ISyncControlDatabaseHelper
-import org.apptank.horus.client.database.IOperationDatabaseHelper
+import org.apptank.horus.client.control.helper.ISyncControlDatabaseHelper
+import org.apptank.horus.client.control.helper.IOperationDatabaseHelper
 import org.apptank.horus.client.di.INetworkValidator
 import org.apptank.horus.client.sync.manager.SynchronizatorManager
 import org.apptank.horus.client.sync.manager.SynchronizatorManager.SynchronizationStatus as SyncStatus
